requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==7.2.1
iniconfig==2.0.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.28.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.4.0
plotly==6.0.0
pluggy==1.5.0
protobuf==4.25.6
pyarrow==19.0.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
pyjokes==0.8.3
pyperclip==1.9.0
pytest==8.3.4
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.2
Random-Word==1.0.11
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
streamlit==1.32.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
zipp==3.21.0
